Move util function from __init__.py to utils.py. In __init__.py, the only thing there is now imports for waveforms
