--- src/ripplegw/ripple.py
@@ -1,14 +1,0 @@
-
-class Waveform:
-
-    model: 
-   
-    def __init__(self) -> None:
-        pass
-    
-    def __call__(self, *args: Any, **kwds: Any) -> Any:
-        pass
-        
-    def full_model(self,
-        
-    )
