+++ src/ripplegw/waveforms/IMRPhenomD_utils.py
@@ -4,7 +4,7 @@ import jax.numpy as jnp
 import jax
 
 from ..constants import gt
-from ..typing import Array
+from jaxtyping import Array
 
 from .IMRPhenomD_QNMdata import QNMData_a, QNMData_fRD, QNMData_fdamp
