+++ src/ripplegw/waveforms/IMRPhenomPv2.py
@@ -1,6 +1,6 @@
 import jax
 import jax.numpy as jnp
-from ripplegw import Mc_eta_to_ms
+from ripplegw.util import Mc_eta_to_ms
 
 from ..constants import gt, MSUN
 import numpy as np
