+++ src/ripplegw/waveforms/TaylorF2.py
@@ -4,7 +4,7 @@ This file implements the TaylorF2 waveform, as described in the LALSuite library
 
 import jax.numpy as jnp
 from ..constants import EulerGamma, gt, m_per_Mpc, PI, MRSUN
-from ripplegw import Mc_eta_to_ms, lambda_tildes_to_lambdas
+from ripplegw.util import Mc_eta_to_ms, lambda_tildes_to_lambdas
 from .IMRPhenom_tidal_utils import get_quadparam_octparam
 from jaxtyping import Array, Float, PyTree
 from ripplegw.waveforms.WaveformModel import WaveformModel, Polarization
